refactor: drop BaseBackend, rely on the Backend Protocol
Each backend is now a standalone class implementing the structural
Backend Protocol directly, instead of inheriting no-op defaults from a
BaseBackend parent. Backends are fully self-describing and a third-party
plugin needs no check-sdist import at all.
