+++ src/check_sdist/backends/maturin.py
@@ -4,10 +4,11 @@ __lazy_modules__ = [f"{__spec__.parent}._base", "pathlib", "typing"]
 
 from typing import Any, ClassVar
 
-from ._base import BaseBackend, glob_filter
+from ._base import glob_filter
 
 TYPE_CHECKING = False
 if TYPE_CHECKING:
+    from collections.abc import Iterator
     from pathlib import Path
 
 __all__ = ["MaturinBackend"]
@@ -17,7 +18,7 @@ def __dir__() -> list[str]:
     return __all__
 
 
-class MaturinBackend(BaseBackend):
+class MaturinBackend:
     """SDist knowledge for the maturin build backend."""
 
     build_backends: ClassVar[tuple[str, ...]] = ("maturin",)
@@ -27,3 +28,8 @@ class MaturinBackend(BaseBackend):
     ) -> frozenset[str]:
         exclude = pyproject.get("tool", {}).get("maturin", {}).get("exclude", [])
         return glob_filter(exclude, files, source_dir)
+
+    def sdist_only_ignores(  # pylint: disable=unused-argument
+        self, pyproject: dict[str, Any]
+    ) -> Iterator[str]:
+        yield from ()
